Revamped Cursor with smooth animations and snapping.
Revamped the cursor to include smooth animations to different tools and sizes, and snaps smoothly to interactable objects. As a consequence, the interactable object interface had to be changed slightly... Also, added an "animal size variable to VE_Animal_Base and parented VE_MacroLib to Objects rather than Actors, so that every object can use the macros. As such, anything using these macros (such as VE_EventManager) had to be altered.

Now the cursor checks if it should be in moving, hovering, or default size, so all that needs to be called to set it to those is either set the "IsMoving" bool to true or use the event "SetCursorHover" (which VE_CursorPlayer does both).