Animals Can Now Target Resident Requirements and "Reside"
It gets a random resident requirement (doesn't check for completion status yet) and eats the desired item. Once it completes the requirement, it "resides" (as in changes a bool and sends a log). Sour Hordo needs 2 Cherries and Sweet Mosquitoast needs 2 lemons.

Updated Player Cursor so it doesn't set an animal as the selected object if it isn't a resident.