Animals Now Only Randomly Roam In Garden Area
Press U to increase garden area and I to decrease garden area. Animals will use EQS to randomly choose a location within the bounds of the Garden Nav Area; for the time being, the animals have a bias for forward movement, but should be able to turn around if they get too close to the boundary. You can also still direct animals outside the garden, but they should be able to go back inside so long as they aren't too far away from the garden. THERE IS NO FAILSAFE FOR WHEN AN ANIMAL IS TOO FAR FROM THE GARDEN. WHEN THIS HAPPENS, DIRECT THEM CLOSER TO GET THEM UNSTUCK!