Tasks With Non-Specific Item Type Logic
Tasks that allow for any kind of an item type is now possible. It uses a data asset that has an item name of "anyitemoftype" and has an EVE_ObjectType enum, For now, I only added tasks for any kind of fruit and any kind of seed. Logic also differentiates between normal and rotting items.