Wildcard Logic
When an animal eats the Wildcard card, it gets a random wildcard feature. Once it's a wildcard, it can't be reverted back, and the wildcard shuffle only applies once. Be sure to mark it as a wildcard part, and if it replaces an existing body part, put the original part in the OriginalMesh part.